linux-dpdk: bump target dpdk version to 17.08
Changes:
- Pcap-based virtual devices are renamed to net_pcap.
- DPDK pktio devices cannot be restarted after calling rte_eth_dev_close().
Move rte_eth_dev_close() calls to global terminate from odp_pktio_close().

+++ platform/linux-dpdk/include/odp/api/plat/packet_inlines.h
@@ -118,7 +118,7 @@ static inline odp_pktio_t _odp_packet_input(odp_packet_t pkt)
 /** @internal Inline function @param pkt @return */
 static inline int _odp_packet_num_segs(odp_packet_t pkt)
 {
-	return _odp_pkt_get(pkt, uint8_t, nb_segs);
+	return _odp_pkt_get(pkt, uint16_t, nb_segs);
 }
 
 /** @internal Inline function @param pkt @return */
